Changes in item of commerce created.

+++ app/src/main/java/com/jorgegomezdeveloper/waycommerce/model/Commerce.kt
@@ -22,5 +22,5 @@ data class Commerce(
     val latitude: Double?,
     @SerializedName("longitude")
     val longitude: Double?,
-    var distance: Float?
+    var distance: Int?
 )
